+++ commons/src/test/java/commons/InstructionTest.java
@@ -7,9 +7,9 @@ public class InstructionTest {
 
     @Test
     public void testConstructor() {
-        Recipe recipe = new Recipe(2);
+        Recipe recipe = new Recipe((long) 2L);
         Instruction instruction = new Instruction("flip", 1, recipe);
-        instruction.setInstructionID(3);
+        instruction.setInstructionID((long)3);
 
         assertEquals(3, instruction.getInstructionID());
         assertEquals("flip", instruction.getDescription());
@@ -19,19 +19,19 @@ public class InstructionTest {
 
     @Test
     public void testSetRecipeID() {
-        Recipe recipe = new Recipe(2);
+        Recipe recipe = new Recipe((long)2L);
         Instruction instruction = new Instruction("flip", 1, recipe);
-        instruction.setInstructionID(3);
+        instruction.setInstructionID((long) 3L);
 
-        instruction.getRecipe().setRecipeID(1);
+        instruction.getRecipe().setRecipeID(Long.valueOf(1));
         assertEquals(1, instruction.getRecipe().getRecipeID());
     }
 
     @Test
     public void testSetDescription() {
-        Recipe recipe = new Recipe(2);
+        Recipe recipe = new Recipe(Long.valueOf(2));
         Instruction instruction = new Instruction("flip", 1, recipe);
-        instruction.setInstructionID(3);
+        instruction.setInstructionID(Long.valueOf(3));
 
         instruction.setDescription("bake");
         assertEquals("bake", instruction.getDescription());
@@ -39,9 +39,9 @@ public class InstructionTest {
 
     @Test
     public void testSetOrderNumber() {
-        Recipe recipe = new Recipe(2);
+        Recipe recipe = new Recipe(Long.valueOf(2));
         Instruction instruction = new Instruction("flip", 1, recipe);
-        instruction.setInstructionID(3);
+        instruction.setInstructionID(Long.valueOf(3));
 
         instruction.setOrderNumber(4);
         assertEquals(4, instruction.getOrderNumber());
@@ -49,14 +49,14 @@ public class InstructionTest {
 
     @Test
     public void testEquals() {
-        Recipe recipe1 = new Recipe(2);
-        Recipe recipe2 = new Recipe(2);
+        Recipe recipe1 = new Recipe(Long.valueOf(2));
+        Recipe recipe2 = new Recipe(Long.valueOf(2));
 
         Instruction instruction1 = new Instruction("flip", 1, recipe1);
         Instruction instruction2 = new Instruction("flip", 1, recipe2);
 
-        instruction1.setInstructionID(3);
-        instruction2.setInstructionID(3);
+        instruction1.setInstructionID(Long.valueOf(3));
+        instruction2.setInstructionID(Long.valueOf(3));
 
         assertEquals(instruction1, instruction2);
         assertEquals(instruction1.hashCode(), instruction2.hashCode());
